broke out single character in to components
